Add WebTorrent Desktop telemetry API
Move all WebTorrent Desktop APIs to their own file

Add comments to server/web.js
